// src/app/vac/models/company.ts
export class Company {
    compCode: string;
    nameTh: string;
    nameEn: string;
    addr1: string;
    addr2: string;
    addr3: string;
    zip: string;
    tel: string;
    fax: string;
    email: string;
    webSite: string;
    regId: string;
    regDate: string;
    vatRate: string;
    todayDate: string;
    useSystemDate: string;
    updateBy: string;
    updateDate: string;
    updateTime: string;
    respMsg: string;
}