// src/app/vac/models/branch.ts
export class Branch {

    compCode: string;
    branchCode: string;
    nameTh: string;
    nameEn: string;
    addr1: string;
    addr2: string;
    addr3: string;
    zip: string;
    tel: string;
    fax: string;
    webSite: string;
    regId: string;
    regDate: string;
    vatRate: string;
    updateBy: string;
    updateDate: string;
    updateTime: string;
    respMsg: string;
}